QrCode/requirements.txt
python-socketio[client]
pyzbar
